+++ lib/msal-browser/test/interaction_client/SilentIframeClient.spec.ts
@@ -146,9 +146,9 @@ describe("SilentIframeClient", () => {
                 tokenType: AuthenticationScheme.BEARER,
             };
             jest.spyOn(
-                                        AuthorizeProtocol,
-                                        "getAuthCodeRequestUrl"
-                                    ).mockResolvedValue(testNavUrl);
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
+            ).mockResolvedValue(testNavUrl);
             jest.spyOn(SilentHandler, "monitorIframeForHash").mockResolvedValue(
                 TEST_HASHES.TEST_SUCCESS_CODE_HASH_SILENT
             );
@@ -187,9 +187,9 @@ describe("SilentIframeClient", () => {
 
         it("Errors thrown during token acquisition are cached for telemetry and browserStorage is cleaned", (done) => {
             jest.spyOn(
-                                        AuthorizeProtocol,
-                                        "getAuthCodeRequestUrl"
-                                    ).mockResolvedValue(testNavUrl);
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
+            ).mockResolvedValue(testNavUrl);
             jest.spyOn(SilentHandler, "monitorIframeForHash").mockRejectedValue(
                 createBrowserAuthError(
                     BrowserAuthErrorCodes.monitorWindowTimeout
@@ -225,9 +225,9 @@ describe("SilentIframeClient", () => {
 
         it("Unexpected non-msal errors do not add correlationId and browserStorage is cleaned", (done) => {
             jest.spyOn(
-                                        AuthorizeProtocol,
-                                        "getAuthCodeRequestUrl"
-                                    ).mockResolvedValue(testNavUrl);
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
+            ).mockResolvedValue(testNavUrl);
             const testError = {
                 errorCode: "Unexpected error",
                 errorDesc: "Unexpected error",
@@ -299,9 +299,9 @@ describe("SilentIframeClient", () => {
                 tokenType: AuthenticationScheme.BEARER,
             };
             jest.spyOn(
-                                        AuthorizeProtocol,
-                                        "getAuthCodeRequestUrl"
-                                    ).mockResolvedValue(testNavUrl);
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
+            ).mockResolvedValue(testNavUrl);
             jest.spyOn(SilentHandler, "monitorIframeForHash").mockResolvedValue(
                 TEST_HASHES.TEST_SUCCESS_CODE_HASH_SILENT
             );
@@ -368,9 +368,9 @@ describe("SilentIframeClient", () => {
                 tokenType: AuthenticationScheme.BEARER,
             };
             jest.spyOn(
-                                        AuthorizeProtocol,
-                                        "getAuthCodeRequestUrl"
-                                    ).mockResolvedValue(testNavUrl);
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
+            ).mockResolvedValue(testNavUrl);
             jest.spyOn(SilentHandler, "monitorIframeForHash").mockResolvedValue(
                 TEST_HASHES.TEST_SUCCESS_CODE_HASH_SILENT
             );
@@ -479,9 +479,9 @@ describe("SilentIframeClient", () => {
                 tokenType: AuthenticationScheme.BEARER,
             };
             jest.spyOn(
-                                        AuthorizeProtocol,
-                                        "getAuthCodeRequestUrl"
-                                    ).mockResolvedValue(testNavUrl);
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
+            ).mockResolvedValue(testNavUrl);
             jest.spyOn(SilentHandler, "monitorIframeForHash").mockResolvedValue(
                 TEST_HASHES.TEST_SUCCESS_NATIVE_ACCOUNT_ID_SILENT
             );
@@ -581,9 +581,9 @@ describe("SilentIframeClient", () => {
                 tokenType: AuthenticationScheme.BEARER,
             };
             jest.spyOn(
-                                        AuthorizeProtocol,
-                                        "getAuthCodeRequestUrl"
-                                    ).mockResolvedValue(testNavUrl);
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
+            ).mockResolvedValue(testNavUrl);
             jest.spyOn(SilentHandler, "monitorIframeForHash").mockResolvedValue(
                 TEST_HASHES.TEST_SUCCESS_NATIVE_ACCOUNT_ID_SILENT
             );
@@ -728,9 +728,9 @@ describe("SilentIframeClient", () => {
                 cloudGraphHostName: "",
             };
             jest.spyOn(
-                                        AuthorizeProtocol,
-                                        "getAuthCodeRequestUrl"
-                                    ).mockResolvedValue(testNavUrl);
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
+            ).mockResolvedValue(testNavUrl);
             jest.spyOn(SilentHandler, "monitorIframeForHash").mockResolvedValue(
                 TEST_HASHES.TEST_SUCCESS_CODE_HASH_SILENT
             );
@@ -780,9 +780,9 @@ describe("SilentIframeClient", () => {
                 status: 200,
             };
             jest.spyOn(
-                                        AuthorizeProtocol,
-                                        "getAuthCodeRequestUrl"
-                                    ).mockResolvedValue(testNavUrl);
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
+            ).mockResolvedValue(testNavUrl);
             jest.spyOn(SilentHandler, "monitorIframeForHash").mockResolvedValue(
                 TEST_HASHES.TEST_SUCCESS_CODE_HASH_SILENT
             );
@@ -989,9 +989,9 @@ describe("SilentIframeClient", () => {
                 tokenType: AuthenticationScheme.BEARER,
             };
             jest.spyOn(
-                                        AuthorizeProtocol,
-                                        "getAuthCodeRequestUrl"
-                                    ).mockResolvedValue(testNavUrl);
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
+            ).mockResolvedValue(testNavUrl);
             jest.spyOn(SilentHandler, "monitorIframeForHash").mockResolvedValue(
                 TEST_HASHES.TEST_SUCCESS_CODE_HASH_SILENT
             );
@@ -1099,9 +1099,9 @@ describe("SilentIframeClient", () => {
                 tokenType: AuthenticationScheme.BEARER,
             };
             jest.spyOn(
-                                        AuthorizeProtocol,
-                                        "getAuthCodeRequestUrl"
-                                    ).mockResolvedValue(testNavUrl);
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
+            ).mockResolvedValue(testNavUrl);
             jest.spyOn(SilentHandler, "monitorIframeForHash").mockResolvedValue(
                 TEST_HASHES.TEST_SUCCESS_CODE_HASH_SILENT
             );
