+++ lib/msal-node/src/index.ts
@@ -12,7 +12,6 @@
  * Warning: This set of exports is purely intended to be used by other MSAL libraries, and should be considered potentially unstable. We strongly discourage using them directly, you do so at your own risk.
  * Breaking changes to these APIs will be shipped under a minor version, instead of a major version.
  */
-
 import * as internals from "./internals.js";
 export { internals };
 
@@ -87,7 +86,11 @@ export {
     AuthorizationCodePayload,
     // Response
     AuthenticationResult,
-    ServerAuthorizationCodeResponse,
+    AuthorizeResponse,
+    /**
+     * @deprecated Use AuthorizeResponse instead
+     */
+    AuthorizeResponse as ServerAuthorizationCodeResponse,
     IdTokenClaims,
     // Cache
     AccountInfo,
