+++ lib/msal-node/src/network/ILoopbackClient.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { ServerAuthorizationCodeResponse } from "@azure/msal-common/node";
+import { AuthorizeResponse } from "@azure/msal-common/node";
 
 /**
  * Interface for LoopbackClient allowing to replace the default loopback server with a custom implementation.
@@ -13,7 +13,7 @@ export interface ILoopbackClient {
     listenForAuthCode(
         successTemplate?: string,
         errorTemplate?: string
-    ): Promise<ServerAuthorizationCodeResponse>;
+    ): Promise<AuthorizeResponse>;
     getRedirectUri(): string;
     closeServer(): void;
 }
