+++ lib/msal-node/src/network/LoopbackClient.ts
@@ -5,7 +5,7 @@
 
 import {
     Constants as CommonConstants,
-    ServerAuthorizationCodeResponse,
+    AuthorizeResponse,
     HttpStatus,
     UrlUtils,
 } from "@azure/msal-common/node";
@@ -26,57 +26,54 @@ export class LoopbackClient implements ILoopbackClient {
     async listenForAuthCode(
         successTemplate?: string,
         errorTemplate?: string
-    ): Promise<ServerAuthorizationCodeResponse> {
+    ): Promise<AuthorizeResponse> {
         if (this.server) {
             throw NodeAuthError.createLoopbackServerAlreadyExistsError();
         }
 
-        return new Promise<ServerAuthorizationCodeResponse>(
-            (resolve, reject) => {
-                this.server = http.createServer(
-                    (req: http.IncomingMessage, res: http.ServerResponse) => {
-                        const url = req.url;
-                        if (!url) {
-                            res.end(
-                                errorTemplate ||
-                                    "Error occurred loading redirectUrl"
-                            );
-                            reject(
-                                NodeAuthError.createUnableToLoadRedirectUrlError()
-                            );
-                            return;
-                        } else if (url === CommonConstants.FORWARD_SLASH) {
-                            res.end(
-                                successTemplate ||
-                                    "Auth code was successfully acquired. You can close this window now."
-                            );
-                            return;
-                        }
-
-                        const redirectUri = this.getRedirectUri();
-                        const parsedUrl = new URL(url, redirectUri);
-                        const authCodeResponse =
-                            UrlUtils.getDeserializedResponse(
-                                parsedUrl.search
-                            ) || {};
-                        if (authCodeResponse.code) {
-                            res.writeHead(HttpStatus.REDIRECT, {
-                                location: redirectUri,
-                            }); // Prevent auth code from being saved in the browser history
-                            res.end();
-                        }
-                        if (authCodeResponse.error) {
-                            res.end(
-                                errorTemplate ||
-                                    `Error occurred: ${authCodeResponse.error}`
-                            );
-                        }
-                        resolve(authCodeResponse);
+        return new Promise<AuthorizeResponse>((resolve, reject) => {
+            this.server = http.createServer(
+                (req: http.IncomingMessage, res: http.ServerResponse) => {
+                    const url = req.url;
+                    if (!url) {
+                        res.end(
+                            errorTemplate ||
+                                "Error occurred loading redirectUrl"
+                        );
+                        reject(
+                            NodeAuthError.createUnableToLoadRedirectUrlError()
+                        );
+                        return;
+                    } else if (url === CommonConstants.FORWARD_SLASH) {
+                        res.end(
+                            successTemplate ||
+                                "Auth code was successfully acquired. You can close this window now."
+                        );
+                        return;
                     }
-                );
-                this.server.listen(0, "127.0.0.1"); // Listen on any available port
-            }
-        );
+
+                    const redirectUri = this.getRedirectUri();
+                    const parsedUrl = new URL(url, redirectUri);
+                    const authCodeResponse =
+                        UrlUtils.getDeserializedResponse(parsedUrl.search) ||
+                        {};
+                    if (authCodeResponse.code) {
+                        res.writeHead(HttpStatus.REDIRECT, {
+                            location: redirectUri,
+                        }); // Prevent auth code from being saved in the browser history
+                        res.end();
+                    }
+                    if (authCodeResponse.error) {
+                        res.end(
+                            errorTemplate ||
+                                `Error occurred: ${authCodeResponse.error}`
+                        );
+                    }
+                    resolve(authCodeResponse);
+                }
+            );
+            this.server.listen(0, "127.0.0.1"); // Listen on any available port
+        });
     }
 
     /**
